+++ src/strojure/ring_undertow/impl/response.clj
@@ -100,7 +100,7 @@
   (send-response-fn
     [_]
     (fn [exchange]
-      (.endExchange exchange))))
+      (.endExchange ^HttpServerExchange exchange))))
 
 ;;,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,