+++ CHANGELOG.md
@@ -8,7 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 Release date `UNRELEASED`
 
-
+- [ impl ] - Handle nil :body - Add type hint to fix reflection warning.
 
 ## `1.0.81`