+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 Release date `UNRELEASED`
 
 - [ impl ] - Handle nil :body - Add type hint to fix reflection warning.
+- [ test ] - Add tests for request helper functions.
 
 ## `1.0.81`